requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
requests==2.31.0
openai==1.3.5
anthropic==0.7.1
nltk==3.8.1
regex==2023.12.25
